GUI can now display the squares in the game
Added the draw_game_state() method, which calls a bunch of helper methods.
Only getting the outermost square to get an outline on mouse hover, even when
hovering over the inside squares, was the annoying. Used canvas.bind_tags()
with a special tag for squares on the outside.

Note: lambda expressions don't have their own enclosed namespace, so using them
as callbacks will use the value of the variable at calltime, not at definition.
To fix, must use default variable.